LayoutTests/fast/performance/performance-now-crash-on-navigated-window-expected.txt

Tests how the performance object works when it's owning window is not fully active due to navigation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS value is not undefined
PASS successfullyParsed is true

TEST COMPLETE
